Add Components

+++ res/Slib/apps/permis/PermisApp.php
@@ -1,15 +1,37 @@
 <?php
 
 class PermisApp extends \Sphp\tools\BasicApp {       
+    /**
+     * $extra array as $extra[table][field] = value format. pass $extra value while 
+     * insert,update data to database. this is extra data which is not provided by bound Components.
+     * @var array
+     */
     public $extra = array();
+    /**
+     * Set Last inserted record "id" value in database when insert operation process.
+     * @var int
+     */
     protected $insertedid = 0;
+    /**
+     * pass default WHERE login to Pagination Component to Search Records. It will only work when 
+     * no other WHERE logic exists.
+     * @var string
+     */
     protected $defWhere = "";
     /**
+     * Form Component ID to process Database bound Component for automatic INSERT,UPDATE and VIEW 
+     * Database record.
+     * @var string default value is form2
+     */
+    protected $formid = "form2";
+    /**
+     * This Front File provide Edit Form to use Insert and Update records
      *  @var \Sphp\tools\FrontFile
      */
     protected $genFormFront = null;
 
     /**
+     * This Front File provide List Records with Pagination + search Form
      *  @var \Sphp\tools\FrontFile
      */
     protected $showallFront = null;
@@ -40,8 +62,9 @@ class PermisApp extends \Sphp\tools\BasicApp {
     
     public function page_event_addnew($param) {
         if($this->page->hasPermission("add")){
-            $this->Client->session("formType", "Add");
-            $this->Client->session("formButton", "Save");
+            $this->genFormFront->addMetaData("pageName", "Add " . $this->genFormFront->getMetaData('pageName'));
+        // set Submit Button Text
+            $this->genFormFront->addMetaData("formButton", "Save");
             $this->setFrontFile($this->genFormFront);
         } else {
             $this->page_new();
@@ -49,11 +72,14 @@ class PermisApp extends \Sphp\tools\BasicApp {
     }
     
     
-    public function page_view() {
-        if($this->page->hasPermission("view")){
-            $this->Client->session("formType", "Edit");
-            $this->Client->session("formButton", "Update");
-            $this->page->viewData($this->genFormFront->getComponent('form2'));
+    public function page_view() { 
+        if($this->page->hasPermission("view") && $this->page->getEventParameter() != ""){
+            $this->genFormFront->addMetaData("pageName", "Edit " . $this->genFormFront->getMetaData('pageName'));
+        // set Submit Button Text
+            $this->genFormFront->addMetaData("formButton", "Update");
+            $form2 = $this->genFormFront->getComponent($this->formid);
+            // fill Components data from database
+            $this->page->viewData($form2);
             $this->setFrontFile($this->genFormFront);
         }else{
             $this->page_new();
@@ -63,6 +89,7 @@ class PermisApp extends \Sphp\tools\BasicApp {
     public function page_insert() {
         global $cmpid;
         if($this->page->hasPermission("add")){
+            if($this->Client->isSession('sid')) setErr('App','Need to login');
             $this->extra[]['userid'] = $this->Client->session('sid');
             $this->extra[]['parentid'] = $this->Client->session('parentid');
             $this->extra[]['spcmpid'] = $cmpid;
@@ -71,16 +98,17 @@ class PermisApp extends \Sphp\tools\BasicApp {
             
             //$this->debug->println("Call Insert Event");
             if (!getCheckErr()) {
-                $this->insertedid = $this->page->insertData($this->extra);            
+                $form2 = $this->genFormFront->getComponent($this->formid);
+                $this->insertedid = $this->page->insertData($form2,$this->extra);            
                 if (!getCheckErr()) {
-                    setMsg('app1', 'Added Successfully');
+                    setMsg('App', 'Added Successfully');
                     $this->setFrontFile($this->showallFront);
                 } else {
-                    setErr('app1', 'Can not add Data');
+                    setErr('App', 'Can not add Data');
                     $this->setFrontFile($this->genFormFront);
                 }
             } else {
-                setErr('app1', 'Can not add Data');
+                setErr('App', 'Can not add Data');
                 $this->setFrontFile($this->genFormFront);
             } 
         } else {
@@ -91,16 +119,17 @@ class PermisApp extends \Sphp\tools\BasicApp {
     public function page_update() {
         if($this->page->hasPermission("view")){
             if (!getCheckErr()) {                        
-                $this->page->updateData($this->extra);
+                $form2 = $this->genFormFront->getComponent($this->formid);
+                $this->page->updateData($form2,$this->extra);
                 if (!getCheckErr()) {
-                    setMsg('app1', 'Update Successfully');
+                    setMsg('App', 'Update Successfully');
                     $this->setFrontFile($this->showallFront);
                 } else {
-                    setErr('app1', 'Record can not update');
+                    setErr('App', 'Record can not update');
                     $this->setFrontFile($this->genFormFront);
                 }
             } else {
-                setErr('app1', 'Record can not update');
+                setErr('App', 'Record can not update');
                 $this->setFrontFile($this->genFormFront);
             }
         } else {
@@ -109,10 +138,10 @@ class PermisApp extends \Sphp\tools\BasicApp {
     }
     
     public function page_delete() {
-        if($this->page->hasPermission("delete")){
+        if($this->page->hasPermission("delete") && $this->page->getEventParameter() != ""){
             $this->page->deleteRec();
             if (!getCheckErr()) {
-                setMsg("app1",'delete Successfully');
+                setMsg("App",'delete Successfully');
                 $type = "danger";                
                 $showall = $this->showallFront->getComponent('showall');
                 $showall->fu_unsetRenderTag();
@@ -120,7 +149,7 @@ class PermisApp extends \Sphp\tools\BasicApp {
                 $this->JSServer->addJSONBlock('html','pagebar',$showall->getPageBar());
                 $this->JSServer->addJSONJSBlock('setFormStatus("'.$msg.'", "'.$type.'"); readyFormAsNew("form2");');            
             } else {
-                setErr('app1', 'Record could not be deleted');
+                setErr('App', 'Record could not be deleted');
                 $this->setFrontFile($this->showallFront);
             }
         } else {
